\name{model}
\alias{model}
\docType{data}
\title{Linear regression model for acceptance of Fluorescence data}
\description{
This data set is the output of the glmnet R package when presented with
a sample training set of acceptable and unacceptable input samples.  The
samples are characterized in terms of average, minimum and maximum 
fluorescence and dfluorescence as well as other measures using the 
statDivMelt function in the DivMelt package.
}
\format{An R formatted version of the glmnet model output.}
\keyword{datasets}

