\name{training}
\alias{training}
\docType{data}
\title{Training data for Linear regression model}
\description{
This data represents a the training set used to generate the "model"
Linear regression model used by default with the DivMelt package.
The training set contains sets of acceptable and unacceptable input samples.
The samples are characterized in terms of average, minimum and maximum 
fluorescence and dfluorescence as well as other measures.  It was generated
using the statDivMelt function in the DivMelt package.  This file serves as input 
to the glmnet function located in the R package by that same name.
}
\format{CSV}
\keyword{datasets}

